pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

[project]
name = "fenicsx_plotly"
version = "0.1.0"
description = "Lightweight library for plotting FEniCS x objects using plotly"
authors = [{name = "Henrik Finsberg", email = "henriknf@simula.no"}]
license = {file = "LICENSE"}
readme = "README.md"
dependencies = ["fenics-dolfinx", "numpy", "plotly", "jupyterlab-widgets"]


[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]
dev = [
    "pdbpp",
    "ipython",

    "pre-commit",
]
docs = [
   "jupyter-book",
   "jupytext",
   "ipython<8.7.0"
]
all = [
   "fenicsx_plotly[test]",
   "fenicsx_plotly[dev]",
   "fenicsx_plotly[docs]"
]

[tool.pytest.ini_options]
addopts = [
   "--import-mode=importlib",
   "--cov=fenicsx_plotly",
   "--cov-report=html",
   "--cov-report=term-missing",
   "-v"
]
testpaths = [
    "tests"
]
